Code refactored into corresponding directories
