+++ app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 27
     defaultConfig {
-        applicationId "com.create.sidhu.movbox"
+        applicationId 'com.roninaks.cinemaclub'
         minSdkVersion 21
         targetSdkVersion 27
         versionCode 1
@@ -17,23 +17,23 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    productFlavors {
+    }
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
     //noinspection GradleCompatible
     implementation 'com.android.support:appcompat-v7:27.1.1'
     //noinspection GradleCompatible
     implementation 'com.android.support:design:27.1.0'
     implementation 'org.apache.httpcomponents:httpclient-android:4.3.5.1'
     implementation 'com.android.support:multidex:1.0.1'
-
     implementation 'com.android.support.constraint:constraint-layout:1.1.2'
     implementation 'com.android.support:support-v4:27.1.1'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-
     //nav-animation dependencies
     implementation 'com.github.ittianyu:BottomNavigationViewEx:1.2.4'
     //circular image view
