+++ app/src/main/java/com/create/sidhu/movbox/fragments/HomeFragment.java
@@ -1,21 +1,22 @@
-package com.create.sidhu.movbox;
+package com.create.sidhu.movbox.fragments;
 
 
 import android.os.Bundle;
-import android.support.design.widget.BottomSheetDialogFragment;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.create.sidhu.movbox.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class PostStatusFragment extends BottomSheetDialogFragment {
+public class HomeFragment extends Fragment {
 
 
-    public PostStatusFragment() {
+    public HomeFragment() {
         // Required empty public constructor
     }
 
@@ -23,8 +24,8 @@ public class PostStatusFragment extends BottomSheetDialogFragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View v= inflater.inflate(R.layout.post_status_dialog,container,false);
-        return v;
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home, container, false);
     }
 
 }
