+++ app/src/main/java/com/create/sidhu/movbox/fragments/PostStatusFragment.java
@@ -1,20 +1,23 @@
-package com.create.sidhu.movbox;
+package com.create.sidhu.movbox.fragments;
 
 
 import android.os.Bundle;
+import android.support.design.widget.BottomSheetDialogFragment;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.create.sidhu.movbox.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class HomeFragment extends Fragment {
+public class PostStatusFragment extends BottomSheetDialogFragment {
 
 
-    public HomeFragment() {
+    public PostStatusFragment() {
         // Required empty public constructor
     }
 
@@ -22,8 +25,8 @@ public class HomeFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false);
+        View v= inflater.inflate(R.layout.post_status_dialog,container,false);
+        return v;
     }
 
 }
