+++ app/src/main/res/layout/fragment_favourites.xml
@@ -1,73 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FavouritesFragment">
-
-
-    <RelativeLayout
+    tools:context=".fragments.FavouritesFragment">
+    <android.support.v7.widget.RecyclerView
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Today"
-            android:textColor="@color/colorTextPrimary"
-            android:layout_margin="5dp"
-            android:textStyle="bold"
-            android:id="@+id/text_today"/>
-
-        <android.support.v7.widget.RecyclerView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/text_today"
-            android:layout_margin="5dp"
-            android:id="@+id/list_view_today">
-        </android.support.v7.widget.RecyclerView>
-
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Yesterday"
-            android:layout_margin="5dp"
-            android:textStyle="bold"
-            android:textColor="@color/colorTextPrimary"
-            android:layout_below="@+id/list_view_today"
-            android:id="@+id/text_yesterday"/>
-
-        <android.support.v7.widget.RecyclerView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/text_yesterday"
-            android:layout_margin="5dp"
-            android:id="@+id/list_view_yesterday">
-        </android.support.v7.widget.RecyclerView>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Earlier"
-            android:layout_margin="5dp"
-            android:textStyle="bold"
-            android:textColor="@color/colorTextPrimary"
-            android:layout_below="@+id/list_view_yesterday"
-            android:id="@+id/text_earlier"/>
-
-        <android.support.v7.widget.RecyclerView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/text_earlier"
-            android:layout_margin="5dp"
-            android:id="@+id/list_view_earlier">
-        </android.support.v7.widget.RecyclerView>
-
-    </RelativeLayout>
-
-
-    <!-- TODO: Update blank fragment layout -->
-
-
-</ScrollView>
+        android:layout_height="wrap_content"
+        android:id="@+id/recyclerView_Favourites">
+    </android.support.v7.widget.RecyclerView>
+</LinearLayout>
