+++ app/src/main/res/layout/fragment_post_status.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PostStatusFragment">
+    tools:context=".fragments.PostStatusFragment">
 
     <RelativeLayout
         android:layout_width="match_parent"
